pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email_summarizer"
version = "0.2.0"
description = "CLI tool to summarize and classify email/LinkedIn threads"
readme = "README.md"
authors = [
  { name="Rohit Rajesh", email="rohitaryan296@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "openai",
  "python-dotenv",
  "google-api-python-client",
  "google-auth-httplib2",
  "google-auth-oauthlib"
]

[project.scripts]
email-summarizer = "email_summarizer.cli:main"
